Another Waystone
Compound nodes are successively introduced, which is able to (or it looks like that it is able to) generate compounds of input symbols (mostly I will not call them "symbols" since there are too many of them).

An evidence is that we can see a clear division between the compound layer and these random layers.

Next I will try this structure on images (MNIST, CIFAR, etc...) Hope I can get a good result!