3rd waystone
I have applied the mechanism on images, and it looks good, but something is wrong with the pattern.

I may think the very beginning definition of these patterns, "the sensation salience". So, I decide to include a part (say 10%) of the input as the pattern. But I also need to consider adding HEAVY noise.